Fix warnings

+++ include/crits.hpp
@@ -7,7 +7,7 @@ namespace criticals
 extern settings::Boolean enabled;
 extern settings::Boolean melee;
 extern std::map<int, std::vector<int>> crit_cmds;
-extern int current_index;
+extern size_t current_index;
 extern bool isEnabled();
 extern bool force_crit_this_tick;
 void fixBucket(IClientEntity *weapon, CUserCmd *cmd);
